# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchSaveRequest": {
        "additionalProperties": false,
        "properties": {
          "urls": {
            "items": {
              "maxLength": 2048,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Urls",
            "type": "array"
          },
          "why_saved": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Saved"
          }
        },
        "required": [
          "urls"
        ],
        "title": "BatchSaveRequest",
        "type": "object"
      },
...
      "DispatchResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "item_public_id": {
            "title": "Item Public Id",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "public_id",
          "item_public_id",
          "attempt",
          "state",
          "error_code",
          "created_at",
          "updated_at"
        ],
        "title": "DispatchResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LibraryItemResponse": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "available_actions": {
            "items": {
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            },
            "title": "Chapters",
            "type": "array"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          },
          "latest_dispatch_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Dispatch Public Id"
          },
          "lifecycle": {
            "enum": [
              "archived",
              "ready",
              "needs_action",
              "failed",
              "processing",
              "queued"
            ],
            "title": "Lifecycle",
            "type": "string"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "saved_at": {
            "format": "date-time",
            "title": "Saved At",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "text_source": {
            "title": "Text Source",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "why_saved": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Saved"
          }
        },
        "required": [
          "public_id",
          "platform",
          "kind",
          "url",
          "title",
          "author",
          "published_at",
          "duration_sec",
          "lang",
          "description",
          "tags",
          "chapters",
          "cover_url",
          "saved_at",
          "why_saved",
          "text_source",
          "lifecycle",
          "error_code",
          "available_actions",
          "latest_dispatch_public_id"
        ],
        "title": "LibraryItemResponse",
        "type": "object"
      },
      "LibraryPageResponse": {
        "additionalProperties": false,
        "properties": {
          "is_true_first_empty": {
            "title": "Is True First Empty",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LibraryItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "is_true_first_empty"
        ],
        "title": "LibraryPageResponse",
        "type": "object"
      },
...
      "TranscriptPageResponse": {
        "additionalProperties": false,
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/TranscriptBlockResponse"
            },
            "title": "Blocks",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "blocks",
          "next_cursor"
        ],
        "title": "TranscriptPageResponse",
        "type": "object"
      },
      "WhySavedRequest": {
        "additionalProperties": false,
        "properties": {
          "why_saved": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Saved"
          }
        },
        "title": "WhySavedRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BrowserSecret": {
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-kb_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Notebook Agent Web API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingest-dispatches/{dispatch_public_id}": {
      "get": {
        "operationId": "dispatch_api_v1_ingest_dispatches__dispatch_public_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dispatch_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Dispatch Public Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Dispatch"
      }
    },
    "/api/v1/library/items": {
      "get": {
        "operationId": "list_items_api_v1_library_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "archived",
                    "ready",
                    "needs_action",
                    "failed",
                    "processing",
                    "queued"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "saved_desc",
              "enum": [
                "saved_desc",
                "saved_asc",
                "title_asc"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Items"
      }
    },
    "/api/v1/library/items/{item_public_id}": {
      "get": {
        "operationId": "detail_api_v1_library_items__item_public_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Item Public Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Detail"
      },
      "patch": {
        "operationId": "update_api_v1_library_items__item_public_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Item Public Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhySavedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Update"
      }
    },
    "/api/v1/library/items/{item_public_id}/transcript": {
      "get": {
        "operationId": "transcript_page_api_v1_library_items__item_public_id__transcript_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Item Public Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Transcript Page"
      }
    },
    "/api/v1/library/items/{item_public_id}:archive": {
      "post": {
        "operationId": "archive_api_v1_library_items__item_public_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Item Public Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Archive"
      }
    },
    "/api/v1/library/items/{item_public_id}:restore": {
      "post": {
        "operationId": "restore_api_v1_library_items__item_public_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Item Public Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Restore"
      }
    },
    "/api/v1/library/items/{item_public_id}:retry": {
      "post": {
        "operationId": "retry_api_v1_library_items__item_public_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_public_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Item Public Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Retry"
      }
    },
    "/api/v1/library/items:batch": {
      "post": {
        "operationId": "save_batch_api_v1_library_items_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSaveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Save Batch"
      }
    }
  }
}